moved function to functions folder
