+++ src/pages/App/App.tsx
@@ -6,18 +6,9 @@ import ReportForm from '../../components/ReportForm/ReportForm';
 import LegalDisclaimer from '../../components/LegalDisclaimer/LegalDisclaimer';
 import UtilButton from '../../components/UtilButton/UtilButton';
 import UtilModal from '../../components/UtilModal/UtilModal';
+import { highlightElement } from '../../functions/uiUtils';
 import Backdrop from '../../components/Backdrop/Backdrop';
 import './App.css';
-
-export function highlightElement(id: string) {
-  const element = document.getElementById(id);
-  if (element) {
-    element.classList.add('highlight');
-    setTimeout(() => {
-      element.classList.remove('highlight');
-    }, 3000);
-  }
-}
 
 function App() {
   const [isReportFormOpen, setIsReportFormOpen] = useState(false);
